QA script 1
This script would input random values to the payeeID to check to see if it would cause issues. This test would test the durability and relability of the system as it would both verify through the url change as well as through a search bar typer. By doing this it allows for the script to verify if there is a mismatch between both values and then therefore would give an error