Rename QA Script 2.py to QA Script2.py
For this script it would make it so that there would be a list of payor names which it would scroll through and search for within the searchbar. By doing this it would allow for the system to more finely go into detail if the payor name is causing issues within the website. The test case in this one is more focused on the Payor names as it would try to register each on and see if there are any mishaps with the results of what the search is giving