Rename QA Script 3.py to QA Script3.py
For this script the test case is the payeeID, it is similar to the second script, but instead of looking for payor names it focuses to see if the payeeID is valid through the search bar as well as the direct link. If there is an issue the script would give a warning of the specific one and then stop.